Migrate VRL CLI into main Rust trunk
