--- lib/remap-cli/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "remap-cli"
-version = "0.1.0"
-authors = ["Vector Contributors <vector@timber.io>"]
-edition = "2018"
-publish = false
-
-[[bin]]
-name = "vrl"
-path = "src/main.rs"
-
-[dependencies]
-bytes = "0.5.6"
-remap = { package = "remap-lang", path = "../remap-lang" }
-remap-functions = { path = "../remap-functions" }
-serde_json = "1"
-structopt = { version = "0.3", default-features = false }
-thiserror = "1"
-rustyline = { version = "7.0.0", default-features = false, optional = true }
-
-[features]
-default = ["repl"]
-repl = ["rustyline"]
