+++ src/vrl/mod.rs
@@ -1,0 +1,17 @@
+mod cmd;
+mod repl;
+
+pub use cmd::run;
+pub use cmd::Opts;
+
+#[derive(thiserror::Error, Debug)]
+enum Error {
+    #[error("io error")]
+    Io(#[from] std::io::Error),
+
+    #[error("remap error: {0}")]
+    Remap(#[from] remap::RemapError),
+
+    #[error("json error")]
+    Json(#[from] serde_json::Error),
+}
